+++ frontend/src/constants/PAYMENT_METHODS.js
@@ -1,0 +1,6 @@
+const PaymentMethods = {
+    CASH: 'Cash',
+    CASH_ON_DELIVERY: 'Cash_on_Delivery',
+}
+
+export default PaymentMethods;
