+++ frontend/src/pages/Orders/components/OrderForm/components/OrderProductTable/OrderProductTable.jsx
@@ -13,7 +13,7 @@ import {
 import { useFormikContext } from 'formik';
 import AddIcon from '@material-ui/icons/Add';
 import RemoveIcon from '@material-ui/icons/Remove';
-import DataDisplayUtils from "../../../../utils/DataDisplayUtils";
+import DataDisplayUtils from "../../../../../../utils/DataDisplayUtils";
 
 const useStyles = makeStyles(() => ({
     table: {
