+++ frontend/src/pages/Orders/components/OrderForm/components/PaymentMethodSelect/PaymentMethodSelect.jsx
@@ -5,37 +5,35 @@ import {TextField} from 'formik-material-ui'
 import {Field} from 'formik'
 
 const useStyles = makeStyles((theme) => ({
-  formControl: {
-    margin: theme.spacing(1),
-    maxWidth: 200,
-  },
-  selectEmpty: {
-    marginTop: theme.spacing(2),
-  },
+    formControl: {
+        margin: theme.spacing(1),
+        width: props => props.width,
+        marginRight: '100%'
+    },
+    selectEmpty: {
+        marginTop: theme.spacing(2),
+    },
 }));
 
-export default function ShopSelect ({shops, name, value}) {
-    const classes = useStyles();
+export default function PaymentMethodSelect (props) {
+    const classes = useStyles(props);
+    const {methods, name} = props;
     return (
         <Field
             component={TextField}
             type="text"
             name={name}
-            label="Client*"
+            label="Payment Method*"
             select
             variant="standard"
-            helperText="Select a Shop (Required)"
             margin="normal"
             InputLabelProps={{
                 shrink: true,
             }}
             className={classes.formControl}
         >
-            <MenuItem value="">
-                <em>None</em>
-            </MenuItem>
-            {shops.map((shop) => (
-                <MenuItem key={shop.id} value={shop.id}>{`${shop.name} [${shop.zone}]`}</MenuItem>
+            {methods.map((method) => (
+                <MenuItem key={method} value={method}>{`${method}`}</MenuItem>
             ))}
         </Field>
     )
